Added more unit tests.

+++ GoatFish.net/Models.cs
@@ -4,9 +4,13 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SQLite;
 using System.Dynamic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using System.Runtime.Serialization;
+using System.Runtime.Serialization.Json;
+using System.Web.Script.Serialization;
 
 namespace GoatFish.net
 {
@@ -15,7 +19,6 @@ namespace GoatFish.net
         private static SQLiteConnection _connection;
         private static SQLiteCommand _command;
         private const string ConnectionString = @"Data Source=test.db;Version=3;";
-
         public Models()
         {
             Initialize();
@@ -57,7 +60,6 @@ namespace GoatFish.net
                 var data = new KeyValuePair<string, string>(uuid, ByteArrayToString((byte[]) reader["data"]));
                 return data;
             }
-
             return new KeyValuePair<string, string>("empty", "empty");
         }
 
@@ -73,7 +75,6 @@ namespace GoatFish.net
             {
                 dictionary.Add(reader["uuid"] as string, ByteArrayToString((byte[]) reader["data"]));
             }
-
             return dictionary;
         }
 
@@ -102,6 +103,8 @@ namespace GoatFish.net
 
         public static void Delete(string uuid)
         {
+            if (Find(uuid).Equals(new KeyValuePair<string, string>("empty", "empty"))) throw new GoatFishSyntaxException("Entity does not exist.");
+
             _command = new SQLiteCommand("DELETE FROM Models WHERE uuid = '" + uuid + "'", _connection);
             _command.ExecuteNonQuery();
         }
@@ -120,8 +123,9 @@ namespace GoatFish.net
         {
             return System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();
         }
-    }
 
+    }
+   
     public class GoatFishSyntaxException : Exception
     {
      
@@ -129,5 +133,27 @@ namespace GoatFish.net
         {
             
         }
+    }
+
+
+    public class JSONSerializer
+    {
+        public static string ToJSON<T>(T list)
+        {
+            var jSerializer = new JavaScriptSerializer();
+            return jSerializer.Serialize(list);
+        }
+
+        public static T Deserialize<T>(string json)
+        {
+            var obj = Activator.CreateInstance<T>();
+            var ms = new MemoryStream(Encoding.Unicode.GetBytes(json));
+            var serializer = new DataContractJsonSerializer(obj.GetType());
+            obj = (T)serializer.ReadObject(ms);
+            ms.Close();
+            return obj;
+        }
+
+
     }
 }
